fixup! (PC-8471) feat(offer): hide see more button if the description is small
